Refactor Update script to migrations: Resolves #3.
